// Front-End/src/app/administrador/interfaces/itinerario.interface.ts
// import { Actividad } from "./actividad.interface";

// export interface Itinerario {
//   iditinerario: string;
//   nombre: string;
//   actividades: Actividad[]
//   alt_img?: string;
// }
export interface Itinerario {
  id: string;
  nombre: string;
  dia_1: string;
  actividades_dia_1: string;
  dia_2: string;
  actividades_dia_2: string;
  dia_3: string;
  actividades_dia_3: string;
  dia_4: string;
  actividades_dia_4: string;
  // alt_img?: string;
}